% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/simrec.R
\name{Anz_Er}
\alias{Anz_Er}
\title{Needed Number of Events}
\usage{
Anz_Er(alpha, Gamma, log_HR)
}
\arguments{
\item{alpha}{significance level}

\item{Gamma}{1 - Power}

\item{log_HR}{log(Hazard Ratio) = Treatment Effect}
}
\description{
Calculates the number of Events L needed
}
\examples{
Anz_Er(alpha=0.05, Gamma=0.2, log_HR = log(0.5))
}
\keyword{Hazard}
\keyword{Ratio}

