% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simrec-pkg.R
\docType{package}
\name{simrec-pkg}
\alias{simrec-pkg}
\title{simrec}
\description{
Simulation of recurrent event data for non-constant baseline hazard
(total-time model)
}
\details{
Simulation of recurrent event data for non-constant baseline hazard 
in the total time model with risk-free intervalls and possibly a competing 
event. The simrec package enables to cut the data to an interim data set, 
and provides functionality to plot.
}
\author{
Katharina Ingel, Stella Preussler, Antje Jahn-Eimermacher, Federico Marini

Maintainer: Antje Jahn-Eimermacher \email{jahna@uni-mainz.de}
}
