% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simrec-pkg.R
\docType{package}
\name{simrec-pkg}
\alias{simrec-pkg}
\alias{simrec-pkg-package}
\title{simrec}
\description{
Simulation of recurrent event data for non-constant baseline hazard
(total-time model)
}
\details{
Simulation of recurrent event data for non-constant baseline hazard 
in the total time model with risk-free intervalls and possibly a competing 
event. Possibility to cut the data to an interim data set. Data can be plotted.
}
\author{
Katharina Ingel, 2016-2017;

Maintainer: TBD \email{TBD@uni-mainz.de}
}
